This needs to be a byte

+++ PolarisServer/Models/Character.cs
@@ -23,7 +23,7 @@ namespace PolarisServer.Models
             Female
         }
 
-        public enum ClassType : ushort
+        public enum ClassType : byte
         {
             Hunter = 0,
             Fighter,
